adding junit tests
