+++ pom.xml
@@ -79,6 +79,12 @@
 			<scope>system</scope>
 			<systemPath>${cdevent-java-sdk-jar}</systemPath>
 		</dependency>
+		<dependency>
+			<groupId>org.mockito</groupId>
+			<artifactId>mockito-inline</artifactId>
+			<version>5.1.1</version>
+			<scope>test</scope>
+		</dependency>
 
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
