--- src/main/java/com/ericsson/event/translator/cdevent/CDEventsTranslator.java
@@ -1,136 +1,0 @@
-package com.ericsson.event.translator.cdevent;
-
-import com.ericsson.eiffel.semantics.events.EiffelTestSuiteFinishedEvent;
-import com.ericsson.eiffel.semantics.events.EiffelTestSuiteStartedEvent;
-import com.ericsson.event.translator.Constants;
-import com.ericsson.event.translator.CustomObjectMapper;
-import com.ericsson.event.translator.cdevent.models.CDEventsData;
-import com.fasterxml.jackson.databind.SerializationFeature;
-import dev.cdevents.CDEventEnums;
-import dev.cdevents.CDEventTypes;
-import io.cloudevents.CloudEvent;
-import io.cloudevents.core.message.MessageWriter;
-import io.cloudevents.http.HttpMessageFactory;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpStatus;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-import java.io.OutputStream;
-import java.io.UncheckedIOException;
-import java.net.HttpURLConnection;
-import java.net.URL;
-
-@Slf4j
-@Component
-public class CDEventsTranslator {
-    @Autowired
-    private CustomObjectMapper objectMapper;
-
-    @Value("${cloudevent.broker.url}")
-    private String CLOUDEVENT_BROKER_URL;
-
-    public boolean translateToCDEvent(String eiffelEventJson, String eiffelEventType) throws IOException {
-        CloudEvent cloudEvent = buildCDEvent(eiffelEventJson, eiffelEventType);
-        if (cloudEvent == null) {
-            log.error("Error translating to CDEvent from Eiffel event type {} ", eiffelEventType);
-            return false;
-        }
-        HttpURLConnection httpStatus = sendCloudEvent(cloudEvent);
-        int statusCode = httpStatus.getResponseCode();
-        if(statusCode == HttpStatus.ACCEPTED.value() || statusCode == HttpStatus.OK.value()){
-            log.info("CDEvent {} is published to events-broker URL - {}", cloudEvent.getType(), CLOUDEVENT_BROKER_URL);
-        }else{
-            log.error("Error sending CDEvent to events-broker response is {} ", httpStatus.getResponseCode());
-            return false;
-        }
-        return true;
-    }
-
-    private CloudEvent buildCDEvent(String eiffelEventJson, String eiffelEventType) throws IOException {
-        CloudEvent cloudEvent = null;
-        if(eiffelEventType.equalsIgnoreCase(Constants.EIFFEL_TESTSUITE_STARTED)){
-            EiffelTestSuiteStartedEvent eiffelTestSuiteStartedEvent = objectMapper.readValue(eiffelEventJson, EiffelTestSuiteStartedEvent.class);
-            cloudEvent = createCDTestSuiteStartedEvent(eiffelTestSuiteStartedEvent);
-        } else if(eiffelEventType.equalsIgnoreCase(Constants.EIFFEL_TESTSUITE_FINISHED)){
-            EiffelTestSuiteFinishedEvent eiffelTestSuiteFinishedEvent = objectMapper.readValue(eiffelEventJson, EiffelTestSuiteFinishedEvent.class);
-            cloudEvent = createCDTestSuiteFinishedEvent(eiffelTestSuiteFinishedEvent);
-        }
-        /*
-        TODO: Different CDEvent types can be created here once the corresponding Eiffel events are identified
-         */
-        return cloudEvent;
-    }
-
-    private CloudEvent createCDTestSuiteStartedEvent(EiffelTestSuiteStartedEvent eiffelTestSuiteStartedEvent) throws IOException {
-        log.info("Create TestSuiteStarted CDEvent from Eiffel event - {}", eiffelTestSuiteStartedEvent.getMeta().getType());
-        CDEventsData cdEventsData = new CDEventsData();
-        cdEventsData.setEventId(eiffelTestSuiteStartedEvent.getMeta().getId());
-        cdEventsData.setEventName(eiffelTestSuiteStartedEvent.getMeta().getType().value());
-        objectMapper.configure(SerializationFeature.FAIL_ON_EMPTY_BEANS, false);
-        /*
-        TODO: match the exact CDEvents fields to Eiffel fields once the latest cdevents/sdk-java is available
-         */
-        CloudEvent cloudEvent = CDEventTypes.createTestEvent(CDEventEnums.TestSuiteStartedEventV1.getEventType(), "testSuiteId", eiffelTestSuiteStartedEvent.getData().getName(), "3.0.0", objectMapper.writeValueAsString(cdEventsData));
-        log.info("CDEvent testSuite started Data {} ", cloudEvent.getData());
-        return cloudEvent;
-    }
-
-    public CloudEvent createCDTestSuiteFinishedEvent(EiffelTestSuiteFinishedEvent eiffelTestSuiteFinishedEvent)
-            throws IOException {
-        log.info("Create TestSuiteFinished CDEvent from Eiffel event - {}", eiffelTestSuiteFinishedEvent.getMeta().getType());
-        CDEventsData cdEventsData = new CDEventsData();
-        cdEventsData.setEventId(eiffelTestSuiteFinishedEvent.getMeta().getId());
-        cdEventsData.setEventName(eiffelTestSuiteFinishedEvent.getMeta().getType().value());
-
-        //For eiffel-demo-cdevents purpose, update the artifact details
-        eiffelTestSuiteFinishedEvent.getData().getCustomData().forEach(data -> {
-            if(data.getKey().equalsIgnoreCase("artifactid")){
-                cdEventsData.setArtifactName(data.getValue().toString());
-            }else if(data.getKey().equalsIgnoreCase("artifactname")){
-                cdEventsData.setArtifactId(data.getValue().toString());
-            }
-        });
-        log.info("TestSuiteFinished cdEventData is {} ", cdEventsData.toString());
-        objectMapper.configure(SerializationFeature.FAIL_ON_EMPTY_BEANS, false);
-        /*
-        TODO: match the exact CDEvents fields to Eiffel fields once the latest cdevents/sdk-java is available
-         */
-        CloudEvent cloudEvent = CDEventTypes.createTestEvent(CDEventEnums.TestSuiteFinishedEventV1.getEventType(), "testSuiteId", "poc", "3.0.0", objectMapper.writeValueAsString(cdEventsData));
-        log.info("CDEvent testSuite finished Data {} ", cloudEvent.getData());
-        return cloudEvent;
-    }
-
-    public HttpURLConnection sendCloudEvent(CloudEvent ceToSend) throws IOException {
-        URL url = new URL(CLOUDEVENT_BROKER_URL);
-        HttpURLConnection httpUrlConnection = (HttpURLConnection) url.openConnection();
-        httpUrlConnection.setRequestMethod("POST");
-        httpUrlConnection.setDoOutput(true);
-        httpUrlConnection.setDoInput(true);
-        MessageWriter messageWriter = createMessageWriter(httpUrlConnection);
-        messageWriter.writeBinary(ceToSend);
-
-        return httpUrlConnection;
-    }
-
-    private MessageWriter createMessageWriter(HttpURLConnection httpUrlConnection) {
-        return HttpMessageFactory.createWriter(
-                httpUrlConnection::setRequestProperty,
-                body -> {
-                    try {
-                        if (body != null) {
-                            httpUrlConnection.setRequestProperty("content-length", String.valueOf(body.length));
-                            try (OutputStream outputStream = httpUrlConnection.getOutputStream()) {
-                                outputStream.write(body);
-                            }
-                        } else {
-                            httpUrlConnection.setRequestProperty("content-length", "0");
-                        }
-                    } catch (IOException t) {
-                        throw new UncheckedIOException(t);
-                    }
-                });
-    }
-}
