--- src/test/java/com/ericsson/event/translator/EventTranslatorApplicationTests.java
@@ -1,13 +1,0 @@
-package com.ericsson.event.translator;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class EventTranslatorApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
